CartSidebar part developed

+++ src/ProductSlice.jsx
@@ -7,8 +7,16 @@ export const productSlice = createSlice({
 },
   reducers: {
     addToCart: (state, action) => {
-      state.value.push(action.payload)
-      localStorage.setItem('productData', JSON.stringify(state.value))
+      const existingProduct=state.value.find((item)=>item.id===action.payload.id)
+      console.log(existingProduct);
+      if(existingProduct){
+        existingProduct.quantity=existingProduct.quantity+action.payload.quantity,
+        existingProduct.price=existingProduct.quantity+action.payload.quantity*existingProduct.price,
+        localStorage.setItem('productData', JSON.stringify(state.value))
+      }else{
+        state.value.push(action.payload)
+        localStorage.setItem('productData', JSON.stringify(state.value))
+      }
     },
     removeFromCard: (state, action) => {
       state.value=state.value.filter((item)=>item.id!==action.payload)
